requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.2
django-import-export==3.3.7
djangorestframework==3.15.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
inflection==0.5.1
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupPy==1.14
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
